Added unit test for oauth
