+++ app/package.json
@@ -14,6 +14,7 @@
     "axios": "^1.6.2",
     "dotenv": "^16.3.1",
     "dotenv-webpack": "^8.0.1",
+    "happy-dom": "^12.10.3",
     "openai": "^4.19.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -22,6 +23,7 @@
     "redaxios": "^0.5.1"
   },
   "devDependencies": {
+    "@testing-library/react": "^14.1.2",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@vitejs/plugin-react-swc": "^3.3.2",
